// src/app/Interface/book.interface.ts
import { Document } from "mongoose";

export interface IBook {
  title: string;
  author: string;
  genre:
    | "FICTION"
    | "NON_FICTION"
    | "SCIENCE"
    | "HISTORY"
    | "BIOGRAPHY"
    | "FANTASY";
  isbn: string;
  description?: string;
  copies: number;
  available: boolean;
}
export interface IBookMethods {
  borrowBook: (quantity: number) => Promise<void>;
}
export type BookDocument = Document & IBook & IBookMethods;